+++ android/src/main/kotlin/com/example/oiti_liveness2d/utils/doccore/DocActivity.kt
@@ -13,27 +13,29 @@ class DocActivity(
     private var context: Context,
     private val result: MethodChannel.Result?,
     private val appKey: String?,
-    private val ticket: String?,
+    private val ticket: String? = null,
     private val themeBuilder: Map<String, String?>?,
+    private val environment: String?,
 ) {
 
     fun getIntent(): Intent {
-        if (appKey.isNullOrBlank() || appKey.length < 16 || ticket?.length!! < 16) {
+        if (appKey.isNullOrBlank() || appKey.length < 16) {
             throw InvalidAppKey()
         }
-
-       // val loadingColor = (loadingAppearance?.get("foreground") ?: "#05D758") as String
-       // val loadingBackgroundColor = (loadingAppearance?.get("background") ?: "#FFFFFF") as String
-       // val loadingSize = (loadingAppearance?.get("size") ?: 1) as Int
-       // val loadingType = getLoadingType((loadingAppearance?.get("type") ?: "") as String)
+        val env: Environment = if (environment.equals("PRD")) Environment.PRD else Environment.HML
+        //val loadingSize: Int =  themeBuilder?.get("setLoadingSpinnerScale")?.toInt()?.times(100) ?: 100
 
         return Intent(context, DocumentscopyActivity::class.java).apply {
-            putExtra(DocumentscopyActivity.PARAM_ENDPOINT, "https://comercial.certiface.com.br")
             putExtra(DocumentscopyActivity.PARAM_APP_KEY, appKey)
-            putExtra(DocumentscopyActivity.PARAM_TICKET,ticket )
+            if(ticket?.isNotEmpty() == true) {
+                putExtra(DocumentscopyActivity.PARAM_TICKET, ticket)
+            }
             putExtra(DocumentscopyActivity.PARAM_THEME, getNewTheme())
             putExtra(DocumentscopyActivity.PARAM_HYBRID, true)
-            putExtra(DocumentscopyActivity.PARAM_ENVIRONMENT, Environment.HML)
+            putExtra(DocumentscopyActivity.PARAM_ENVIRONMENT, env)
+            putExtra(DocumentscopyActivity.PARAM_CUSTOM_LOADING_BACKGROUND, themeBuilder?.get("setLoadingBackgroundColor") as? String ?: "#333333")
+            putExtra(DocumentscopyActivity.PARAM_CUSTOM_LOADING_SPINNER_COLOR, themeBuilder?.get("setLoadingSpinnerColor") as? String ?: "#05D758")
+            putExtra(DocumentscopyActivity.PARAM_CUSTOM_LOADING_SIZE, 100)
             putExtra(
                 DocumentscopyActivity.PARAM_DEBUG_ON,
                 false
@@ -41,10 +43,42 @@ class DocActivity(
         }
     }
 
+    private fun getNewTheme() = DocCopyTheme.Builder()
+    .instructionTitleText(themeBuilder?.get("setInstructionTitleText") ?: "Envio de documento")
+    .instructionTitleColor(themeBuilder?.get("setInstructionTitleColor") ?: "#52575D")
+    .instructionCaptionColor(themeBuilder?.get("setInstructionCaptionColor") ?: "#666666")
+    .instructionCaptionText(themeBuilder?.get("setInstructionCaptionText") ?: "Isso garante que seu documento é oficial.")
+    .instructionBottomSheetBackgroundColor(themeBuilder?.get("setInstructionBottomSheetColor")  ?: "#FFFFFF")
+    //.instructionBottomSheetRadius(themeBuilder?.getInt("setInstructionBottomSheetRadius") ?: 20)
+    .setInstructionOptionDocumentBackgroundColor(themeBuilder?.get("setInstructionDocOptionBackgroundColor")  ?: "#F5F5F5")
+    .setInstructionOptionDocumentText(themeBuilder?.get("setInstructionDocOptionTitleText")  ?: "Use RG, CNH, DNI ou CIN.")
+    .setInstructionOptionDocumentTextColor(themeBuilder?.get("setInstructionDocOptionTitleColor")  ?: "#000000")
+    .setInstructionOptionDocumentTextFont(themeBuilder?.get("setInstructionDocOptionTitleFont")  ?: "inter_bold")
+    .setInstructionOptionDocumentBorderColor(themeBuilder?.get("setInstructionDocOptionBorderColor")  ?: "#E0E0E0")
+    .setInstructionOptionLightingBackgroundColor(themeBuilder?.get("setInstructionEnvOptionBackgroundColor")  ?: "#F5F5F5")
+    .setInstructionOptionLightingText(themeBuilder?.get("setInstructionEnvOptionTitleText")  ?: "Retire o documento do plástico de proteção.")
+    .setInstructionOptionLightingTextColor(themeBuilder?.get("setInstructionEnvOptionTitleColor")  ?: "#000000")
+    .setInstructionOptionLightingTextFont(themeBuilder?.get("setInstructionEnvOptionTitleFont")  ?: "inter_bold")
+    .setInstructionOptionLightingBorderColor(themeBuilder?.get("setInstructionEnvOptionBorderColor")  ?: "#E0E0E0")
+    .setInstructionContinueButtonBackgroundColor(themeBuilder?.get("setInstructionContinueButtonBackgroundColor")  ?: "#F5F5F5")
+    .setInstructionContinueButtonTextColor(themeBuilder?.get("setInstructionContinueButtonTextColor")  ?: "#000000")
+    .setInstructionContinueButtonTextFont(themeBuilder?.get("setInstructionContinueButtonFont")  ?: "ubuntu_regular")
 
-    fun onDocumentscopyResultSuccess() {
-        result?.success("Doc - RESULT_OK")
-    }
+    .setCaptureBackgroundColor(themeBuilder?.get("setCaptureBackgroundColor")  ?: "#1E1E1E")
+    .setCaptureInstructionGuideTextFront(themeBuilder?.get("setCaptureInstructionGuideTextFront")  ?: "Posicione a frente do seu documento dentro da marcação em uma superfície plana e fotografe.")
+    .setCaptureInstructionGuideTextBack(themeBuilder?.get("setCaptureInstructionGuideTextBack")  ?: "Posicione o verso do seu documento dentro da marcação em uma superfície plana e fotografe.")
+    .setTextOk(themeBuilder?.get("setCaptureUsePictureButtonConfirmationText")  ?: "Sim")
+    .setTextRedo(themeBuilder?.get("setCaptureTakeNewPictureButtonText")  ?: "Tentar Novamente")
+    .setCaptureInstructionGuideTextColor(themeBuilder?.get("setCaptureInstructionGuideTextColor")  ?: "#FFFFFF")
+    .setTextConfirmation(themeBuilder?.get("setTextConfirmation")  ?: "A foto do documento ficou boa?")
+    .setBackgroundOkColor(themeBuilder?.get("setCaptureUsePictureButtonHighlightedStateColorsBackground")  ?: "#05D758")
+    .setCaptureInstructionGuideReviewText(themeBuilder?.get("setCaptureInstructionGuideReviewText")  ?: "Confirme se os dados ficaram nítidos e os textos legíveis.")
+    .setCapturePreviewBorderColorForCapture(themeBuilder?.get("setCapturePreviewBorderColorForCapture")  ?: "#05D758")
+    .setCaptureBottomSheetShapeColor(themeBuilder?.get("setCaptureBottomSheetShapeColor")  ?: "#FFFFFF")
+    .setCaptureBackButtonColorsIcon(themeBuilder?.get("setCaptureBackButtonColorsIcon")  ?: "#FFFFFF")
+    .setCaptureCloseButtonColorsIcon(themeBuilder?.get("setCaptureBackButtonColorsIcon")  ?: "#FFFFFF")
+    .setBackgroundDismissColor(themeBuilder?.get("setCaptureTakeNewPictureButtonNormalStateColorsBackground")  ?: "#F5F5F5")
+        .build()
 
     fun onDocumentscopyCancelled(data: Intent?) {
         val errorMessage = data?.getStringExtra(DocumentscopyActivity.PARAM_RESULT_ERROR)
@@ -56,74 +90,17 @@ class DocActivity(
         }
     }
 
-    private fun getNewTheme() = DocCopyTheme.Builder()
-        //NewCustom Instruction
-        .instructionTitleText("Aro aqui eu To")
-        .instructionTitleColor("#A5CD39")
-        .instructionCaptionColor("#A5CD39")
-        .instructionCaptionText("123 mais um teste")
-        .instructionBackgroundColor("#A5CD39")
-        .instructionBottomSheetBackgroundColor("#333333")
-        .rgTitleText("title RG")
-        .rgTitleColor("#FFFFFF")
-        .rgCaptionText("RG Subtext Lorem Ipsum")
-        .rgCaptionColor("#FFFFFF")
-        .cnhTitleText("title CNH")
-        .cnhTitleColor("#FFFFFF")
-        .cnhCaptionText("CNH Subtext Lorem Ipsum")
-        .cnhCaptionColor("#FFFFFF")
-
-        //NewCustom Capture
-        .setCaptureBackgroundColor("#FFFFFF")
-        .textOkColor("#FFFFFF")
-        .setBackgroundDismissColor("#DD0101")
-        .setTryAgainColor("#FFFFFF")
-        .setBackgroundOkColor("#4bb75f")
-
-        .setTextFront("textFront")
-        .setTextBack("textBack")
-        .setCaptureInstructionGuideText("textPicture")
-        .setCaptureInstructionGuideBackgroundColor("#DD0101")
-        .setCaptureInstructionGuideTextColor("#333333")
-        .setTextOk("textOk")
-        .setTextConfirmation("textConfirmation")
-        .setTextRedo("textRedo")
-
-        // caso seja hybrido
-        /*.setBackgroundDismissColorHybrid("#73997d")
-        .setBackgroundOkColorHybrid("#3b3bd4")
-        .setTryAgainColorHybrid("#383531")
-        .setBackgroundCameraColorHybrid("#70856e")*/
-
-        .build()
-
-
-    /*
-    fun onDocResultSuccess(data: Intent?) {
+    fun onDocumentscopyResultSuccess(data: Intent?) {
         val response = mapOf<String, Any?>(
-            "valid" to data?.getBooleanExtra(DocumentscopyActivity.PARAM_RESULT_VALID, false),
-            "cause" to data?.getStringExtra(DocumentscopyActivity.PARAM_RESULT_CAUSE),
-            "codId" to data?.getStringExtra(DocumentscopyActivity.PARAM_RESULT_COD_ID),
-            "protocol" to data?.getStringExtra(DocumentscopyActivity.PARAM_RESULT_PROTOCOL),
-            "blob" to data?.getStringExtra(DocumentscopyActivity.PARAM_RESULT_SCAN),
+            "valid" to "true",
+            "cause" to "0",
+            "codId" to "0",
+            "protocol" to "0",
+            "blob" to "0",
         )
 
         result?.success(response)
     }
 
-    fun onLiveness3DResultCancelled(data: Intent?) {
-        val errorMessage: String = data?.getStringExtra(DocumentscopyActivity.PARAM_RESULT_ERROR) ?: ""
-        Log.d("TAG", errorMessage)
-        result?.error(errorMessage, errorMessage, null)
-    }
 
-
-    private fun getLoadingType(typeString: String): LoadingType3D {
-        return when(typeString) {
-            "spinner" -> LoadingType3D.SPINNER
-            "activity" -> LoadingType3D.ACTIVITY_INDICATOR
-            else -> LoadingType3D.ACTIVITY_INDICATOR
-        }
-    }
-     */
 }