+++ android/src/main/kotlin/com/example/oiti_liveness2d/utils/liveness2d/Liveness2dActivity.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import android.content.Intent
 import android.util.Log
 import br.com.oiti.certiface.data.FaceCaptchaErrorCode
+import br.com.oiti.certiface.data.util.Environment
 import br.com.oiti.certiface.facecaptcha.FaceCaptchaActivity
 import br.com.oiti.certiface.facecaptcha.UserData
 import br.com.oiti.certiface.documentscopy.DocumentscopyActivity
@@ -15,9 +16,8 @@ class Liveness2dActivity(
     private var context: Context,
     private val result: MethodChannel.Result?,
     private val appKey: String?,
+    private val environment: String?,
 ) {
-
-    private val ENDPOINT = "https://comercial.certiface.com.br:8443"
 
     fun getIntent(): Intent {
         if (appKey.isNullOrBlank()) {
@@ -25,25 +25,24 @@ class Liveness2dActivity(
         }
 
         val userData = UserData(appKey = appKey)
-
-       // val loadingColor = (loadingAppearance?.get("foreground") ?: "#05D758") as String
-       // val loadingBackgroundColor = (loadingAppearance?.get("background") ?: "#FFFFFF") as String
-       // val loadingSize = (loadingAppearance?.get("size") ?: 1) as Int
+        val env: Environment = if (environment.equals("PRD")) Environment.PRD else Environment.HML
 
         return Intent(context, FaceCaptchaActivity::class.java).apply {
-            putExtra(FaceCaptchaActivity.PARAM_ENDPOINT, ENDPOINT)
-            putExtra(FaceCaptchaActivity.PARAM_USER_DATA, userData)
             putExtra(FaceCaptchaActivity.PARAM_SHOW_INSTRUCTIONS, false)
+            putExtra(FaceCaptchaActivity.PARAM_ENVIRONMENT, env)
+            putExtra(FaceCaptchaActivity.PARAM_USER_DATA, userData)
+            putExtra(FaceCaptchaActivity.PARAM_DEBUG_ON, false)
         }
     }
 
 
     fun onFaceCaptchaResultSuccess(data: Intent?) {
+        val codID = data?.getDoubleExtra(FaceCaptchaActivity.PARAM_RESULT_COD_ID, 0.0)
         val response = mapOf<String, Any?>(
-            "result" to data?.getBooleanExtra(FaceCaptchaActivity.PARAM_RESULT, false),
+            "valid" to data?.getBooleanExtra(FaceCaptchaActivity.PARAM_RESULT, false),
             "cause" to data?.getStringExtra(FaceCaptchaActivity.PARAM_RESULT_CAUSE),
-            "codID" to data?.getDoubleExtra(FaceCaptchaActivity.PARAM_RESULT_COD_ID, 0.0),
-            "protocol" to data?.getStringExtra(FaceCaptchaActivity.PARAM_RESULT_PROTOCOL),
+            "cod_id" to codID.toString(),
+            "uid_protocol" to data?.getStringExtra(FaceCaptchaActivity.PARAM_RESULT_PROTOCOL),
         )
 
         result?.success(response)
@@ -59,21 +58,4 @@ class Liveness2dActivity(
         }
     }
 
-
-    /*
-
-    fun onLiveness3DResultCancelled(data: Intent?) {
-        val errorMessage: String = data?.getStringExtra(HybridLiveness3DActivity.PARAM_RESULT_ERROR) ?: ""
-        Log.d("TAG", errorMessage)
-        result?.error(errorMessage, errorMessage, null)
-    }
-
-    private fun getLoadingType(typeString: String): LoadingType3D {
-        return when(typeString) {
-            "spinner" -> LoadingType3D.SPINNER
-            "activity" -> LoadingType3D.ACTIVITY_INDICATOR
-            else -> LoadingType3D.ACTIVITY_INDICATOR
-        }
-    }
-    */
 }