+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+## 2.1.0
+
+_data: 12/03/2024_
+
+- **Novidades**
+  - Atualizado customização na DocCore.
+
+## 2.0.1
+
+_data: 20/12/2023_
+
+- **Novidades**
+  - Adicionado README.MD
+
+## 2.0.0
+
+_data: 20/12/2023_
+
+- **Novidades**
+  - Customização de textos, cores e fontes;
+  - Telas de Intruções e Permissões via Widget
+  - Import com apenas uma chamada;
+
+## 1.0.0
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+- TODO: Describe initial release.